Memory allocation fix; Add logger config

+++ esphome/components/lvgl/lvgl_hal.h
@@ -11,6 +11,6 @@
 #endif
 
 EXTERNC unsigned long lv_millis(void);
-EXTERNC void * lv_custom_mem_alloc(size_t size);
+EXTERNC void * lv_custom_mem_alloc(unsigned int size);
 EXTERNC void lv_custom_mem_free(void * ptr);
-EXTERNC void * lv_custom_mem_realloc(void * ptr, size_t size);
+EXTERNC void * lv_custom_mem_realloc(void * ptr, unsigned int size);
